RF - rename ClosestDirectionGetter to ClosestPeakDirectionGetter
