+++ dipy/direction/__init__.py
@@ -1,4 +1,6 @@
 
+from .bootstrap_direction_getter import BootDirectionGetter
+from .closest_peak_direction_getter import ClosestPeakDirectionGetter
 from .probabilistic_direction_getter import ProbabilisticDirectionGetter
 from .probabilistic_direction_getter import DeterministicMaximumDirectionGetter
 from .peaks import *
